function-calling: start refactor extract functions into component dirs
