+++ packages/function-calling/src/writeSourceFile/index.ts
@@ -1,0 +1,3 @@
+export * from './writeSourceFile'
+import config from './writeSourceFile'
+export default config
